Working partial Integration for Ethernet + Parsing and SD Card + Web Server
